requirements.txt
appdirs==1.4.3
APScheduler==3.6.3
bcrypt==3.1.3
beautifulsoup4==4.7.1
bs4==0.0.1
cffi==1.10.0
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
Django==1.11.17
django-audit-log==0.7.0
django-auditlog==0.4.3
django-background-tasks==1.1.10
django-compat==1.0.14
django-cors-headers==2.1.0
django-jsonfield==1.0.1
django-rest-swagger==2.2.0
django-user-agents==0.4.0
djangorestframework==3.6.2
drf-extensions==0.3.1
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
Flask==0.12.2
future==0.17.1
geocoder==1.38.1
geoip2==2.9.0
itsdangerous==1.1.0
itypes==1.1.0
Jinja2==2.10.1
linecache2==1.0.0
MarkupSafe==1.1.1
maxminddb==1.4.1
num2words==0.5.7
olefile==0.46
openapi-codec==1.3.2
packaging==16.8
Pillow==4.1.1
psycopg2-binary==2.8.3
pycparser==2.17
PyJWT==1.5.0
pyparsing==2.2.0
PySocks==1.6.7
python-dateutil==2.8.0
pytz==2017.2
PyYAML==5.1.2
ratelim==0.1.6
requests==2.14.2
simplejson==3.16.0
six==1.10.0
slugify==0.0.1
soupsieve==1.7.3
tinys3==0.1.12
traceback2==1.4.0
twilio==6.1.1
tzlocal==2.0.0
ua-parser==0.8.0
Unidecode==1.0.23
unittest2==1.1.0
uritemplate==3.0.0
urllib3==1.21.1
user-agents==2.0
validate-email==1.3
Werkzeug==0.15.5
xlrd==1.1.0
XlsxWriter==1.0.5
xlwt==1.3.0
pycryptodome==3.4.3
pyfcm==1.4.7